Many Rubocop lint corrections (rubocop --lint -D)
Squashed commit of the following:

commit 15fac26b4f3deb5fc6c542d9d7b9e7e0a76f8cdf
Merge: e945939 a8820fa
Author: Ramón Cahenzli <rca@psy-q.ch>
Date:   Wed Apr 29 10:28:19 2015 +0200

    Merge branch 'master' into rca/rubocop

commit e9459395d0cc414d7c9a101785a00dfb4d230b05
Author: Ramón Cahenzli <rca@psy-q.ch>
Date:   Wed Apr 29 10:27:09 2015 +0200

    Rubocop: Some assorted linting.

commit 39551c3cc9995dcfdada17cb7585517a0aaee6ca
Author: Ramón Cahenzli <rca@psy-q.ch>
Date:   Wed Apr 29 08:58:10 2015 +0200

    Rubocop: Trying to fix Lint/EndAlignment without getting sidetracked too much.

commit dd3f343e48e17dc9ee11f0ff12f8f1ef543487c2
Author: Ramón Cahenzli <rca@psy-q.ch>
Date:   Wed Apr 29 08:20:15 2015 +0200

    Rubocop: Authentication controllers are quite bad.

commit aefb49a67a8c1da0df65261425153cd9e00cfe05
Author: Ramón Cahenzli <rca@psy-q.ch>
Date:   Tue Apr 28 16:32:18 2015 +0200

    Fix syntax error in model factory.

commit f77a02e3c04dd5d4e0ac055ae7fc0cb3dcde3b87
Author: Ramón Cahenzli <rca@psy-q.ch>
Date:   Tue Apr 28 16:18:39 2015 +0200

    Fixing model_factory.

commit ae3340d711466d7cca414e69c43b9cb4130b074c
Author: Ramón Cahenzli <rca@psy-q.ch>
Date:   Tue Apr 28 16:02:55 2015 +0200

    Rubocop: Linting leihs_factory and it looks like FactoryGirl doesn't like lambda syntax.

commit 013fc47b7fb832ff949a50a1cef180c58f0a9220
Author: Ramón Cahenzli <rca@psy-q.ch>
Date:   Tue Apr 28 15:48:42 2015 +0200

    Rubocop: Reverting supplier_factory.rb because I don't know how that bit works.

commit 6da19ac993bf6a2a1bc4fd8a5328d25909d5ece9
Author: Ramón Cahenzli <rca@psy-q.ch>
Date:   Tue Apr 28 15:25:12 2015 +0200

    Rubocop: Linting features/reports/routes_report.rb

commit 2d1dae8b15bd146f95599c8aacbf1afd0ae3edbd
Author: Ramón Cahenzli <rca@psy-q.ch>
Date:   Tue Apr 28 12:56:16 2015 +0200

    Rubocop: linting factories/model_factory.rb factories/user_factory.rb features/step_definitions/web_steps.rb

commit 77033a322e44a2bf8a8d68d9dc68e3ec742f45e8
Author: Ramón Cahenzli <rca@psy-q.ch>
Date:   Tue Apr 28 12:48:29 2015 +0200

    Linted factories/supplier_factory.rb

commit d4d11dfb7fd58c1ca106ff47ea1489d3c3a65141
Merge: 3db5503 34eb4c4
Author: Ramón Cahenzli <rca@psy-q.ch>
Date:   Tue Apr 28 11:53:49 2015 +0200

    Conflict resolution.

commit 3db550363e9081c82e24aec77c40018f3b44f8d0
Author: Ramón Cahenzli <rca@psy-q.ch>
Date:   Tue Apr 28 11:16:44 2015 +0200

    Commenting steps.

commit 339eea343958755bc6d82f78e8f9dd25577a6ab6
Author: Ramón Cahenzli <rca@psy-q.ch>
Date:   Tue Apr 28 10:43:16 2015 +0200

    Rubocop: Lint/StringConversionInInterpolation auto-corrected.

commit d33e8c73d9f08946303c1162f9e22be4720dcedc
Author: Ramón Cahenzli <rca@psy-q.ch>
Date:   Tue Apr 28 10:41:50 2015 +0200

    Let's go easy on those line lengths for now.

commit 1cc9d95ef377591b066cccd89a47c6afc08c0e76
Author: Ramón Cahenzli <rca@psy-q.ch>
Date:   Tue Apr 28 10:41:24 2015 +0200

    Rubocop: Mostly linted lib/tasks/app/seed_demo.rake

commit 6b975e8d399e9c1822f88f8e34508e78a01eb3ce
Author: Ramón Cahenzli <rca@psy-q.ch>
Date:   Tue Apr 28 10:39:02 2015 +0200

    Rubocop: Mostly linted lib/tasks/app/po2json.rake

commit ffa4519e15bce373381f9992a0bc245cba0307b9
Author: Ramón Cahenzli <rca@psy-q.ch>
Date:   Tue Apr 28 10:33:11 2015 +0200

    This file is not needed anymore.

commit 032e0f5ae81f59ce476ea1f3bf49936741722c7e
Author: Ramón Cahenzli <rca@psy-q.ch>
Date:   Tue Apr 28 10:27:16 2015 +0200

    Rubocop: Mostly linted features/step_definitions/examples/software_steps.rb

commit 9d8740db888a10174de93d39cb584ae93826fa5c
Author: Ramón Cahenzli <rca@psy-q.ch>
Date:   Tue Apr 28 10:21:01 2015 +0200

    Rubocop: app/models/database_authentication.rb linted a little

commit 10db3c537a4f77a5006a5f78097eef8836e1549f
Author: Ramón Cahenzli <rca@psy-q.ch>
Date:   Tue Apr 28 10:05:53 2015 +0200

    Rubocop: Debugger exceptions.

commit 1aed6822b3e4962e22a4bc0e2315faa78ddec90e
Author: Ramón Cahenzli <rca@psy-q.ch>
Date:   Tue Apr 28 09:53:41 2015 +0200

    Rubocop: Style/StringLiterals,Style/HashSyntax

+++ app/controllers/admin/scenarios_controller.rb
@@ -3,17 +3,18 @@ class Admin::ScenariosController < Admin::ApplicationController
   def index
     dir = File.join(Rails.root, 'features', 'reports')
     merged_file = File.join(dir, 'merged_report.json')
-    @routes_scenarios = if File.exists?(merged_file)
-      JSON.load(File.new(merged_file))
-    else
-      merged_report = Dir.glob(File.join(dir, '*.json')).inject({}) do |h, jsonfile|
-        h.deep_merge JSON.load(File.new(jsonfile))
+    @routes_scenarios =
+      if File.exists?(merged_file)
+        JSON.load(File.new(merged_file))
+      else
+        merged_report = Dir.glob(File.join(dir, '*.json')).inject({}) do |h, jsonfile|
+          h.deep_merge JSON.load(File.new(jsonfile))
+        end
+        File.open(merged_file, 'w') do |f|
+          f.write(merged_report.to_json)
+        end
+        merged_report
       end
-      File.open(merged_file,'w') do |f|
-        f.write(merged_report.to_json)
-      end
-      merged_report
-    end
   end
 
 end
